chore(types): refactoring gocd types
