+++ src/api/gocd/index.test.ts
@@ -1,5 +1,5 @@
 import { GOCD_ORIGIN, GOCD_TOKEN } from '@/config';
-import { GoCDDashboardResponse } from '@/types';
+import { GoCDDashboardResponse } from '@/types/gocd';
 import * as iap from '@/utils/iap';
 
 import { fetchDashboard, removeNestedEmbeddings } from './index';
