+++ src/brain/gocdDataDog/index.ts
@@ -3,7 +3,7 @@ import { gocdevents } from '@/api/gocdevents';
 //   GOCD_SENTRYIO_BE_PIPELINE_NAME,
 //   GOCD_SENTRYIO_FE_PIPELINE_NAME,
 // } from '@/config';
-import { GoCDResponse } from '@/types';
+import { GoCDResponse } from '@/types/gocd';
 
 import { DeployDatadogEvents } from './deployDatadogEvents';
