+++ src/brain/gocdNoDeploysAlert/index.ts
@@ -14,7 +14,7 @@ import {
   GoCDPipeline,
   GoCDResponse,
   GoCDStageResponse,
-} from '@/types';
+} from '@/types/gocd';
 import { getLastGetSentryGoCDDeploy } from '@/utils/db/getLatestDeploy';
 
 const PIPELINE_FILTER = [
